DriveSetup: fix creating partitions
The code paths for creating a new partition and editing an existing one
are more mixed together than I thought:

- CreateParametersPanel reuses a method from ChangeParametersPanel which
  I had modified.
- In the storage kit, the code to create the parameter editor is also
  shared between the two.

The problem is that when creating a partition, we have only the parent
BPartition. When modifying parameters, we have the child, and we need
to get some things from the child (the current parameters), and some
from the parent (the parameter editor).

Fixes #16859.

Change-Id: I5978386978c79a351e94c098c065fd43c422eb51
Reviewed-on: https://review.haiku-os.org/c/haiku/+/3822
Reviewed-by: Adrien Destugues <pulkomandy@gmail.com>