sdhci: actually wait for commands to complete
The code allowed execution to continue as soon as the "command inhibit"
bit was cleared. This is incorrect: we need to wait for a command result
(either command complete, or timeout) to be available before continuing.

This should fix #17031.

Change-Id: I8f3fe60c2e47582b399952b19c05c6ed2161afd7
Reviewed-on: https://review.haiku-os.org/c/haiku/+/4121
Reviewed-by: Jérôme Duval <jerome.duval@gmail.com>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
(cherry picked from commit ab9943450324a0401ec2e9a0030de6a17347f51b)
Reviewed-on: https://review.haiku-os.org/c/haiku/+/4023
Reviewed-by: Alex von Gluck IV <kallisti5@unixzen.com>